+++ material-stepper/src/main/java/com/stepstone/stepper/internal/feedback/StepperFeedbackTypeFactory.java
@@ -16,6 +16,7 @@ limitations under the License.
 
 package com.stepstone.stepper.internal.feedback;
 
+import android.support.annotation.NonNull;
 import android.support.annotation.RestrictTo;
 
 import com.stepstone.stepper.StepperLayout;
@@ -26,17 +27,20 @@ import static android.support.annotation.RestrictTo.Scope.LIBRARY;
  * Factory class for creating feedback stepper types.
  */
 @RestrictTo(LIBRARY)
-public class StepperFeedbackTypeFactory {
+public final class StepperFeedbackTypeFactory {
+
+    private StepperFeedbackTypeFactory() {}
 
     /**
      * Creates a stepper feedback type for provided arguments.
      * It can be a composition of several feedback types depending on the provided flags.
      *
      * @param feedbackTypeMask step feedback type mask, should contain one or more from {@link StepperFeedbackType}
-     * @param stepperLayout stepper layout to use with the chosen stepper feedback type(s)
+     * @param stepperLayout    stepper layout to use with the chosen stepper feedback type(s)
      * @return a stepper feedback type
      */
-    public static StepperFeedbackType createType(int feedbackTypeMask, StepperLayout stepperLayout) {
+    @NonNull
+    public static StepperFeedbackType createType(int feedbackTypeMask, @NonNull StepperLayout stepperLayout) {
 
         StepperFeedbackTypeComposite stepperFeedbackTypeComposite = new StepperFeedbackTypeComposite();
 
@@ -49,12 +53,24 @@ public class StepperFeedbackTypeFactory {
             stepperFeedbackTypeComposite.addComponent(new TabsStepperFeedbackType(stepperLayout));
         }
 
-        if ((feedbackTypeMask & StepperFeedbackType.CONTENT) != 0) {
-            stepperFeedbackTypeComposite.addComponent(new ContentStepperFeedbackType(stepperLayout));
+        if ((feedbackTypeMask & StepperFeedbackType.CONTENT_PROGRESS) != 0) {
+            stepperFeedbackTypeComposite.addComponent(new ContentProgressStepperFeedbackType(stepperLayout));
+        }
+
+        if ((feedbackTypeMask & StepperFeedbackType.CONTENT_FADE) != 0) {
+            stepperFeedbackTypeComposite.addComponent(new ContentFadeStepperFeedbackType(stepperLayout));
+        }
+
+        if ((feedbackTypeMask & StepperFeedbackType.CONTENT_OVERLAY) != 0) {
+            stepperFeedbackTypeComposite.addComponent(new ContentOverlayStepperFeedbackType(stepperLayout));
         }
 
         if ((feedbackTypeMask & StepperFeedbackType.DISABLED_BOTTOM_NAVIGATION) != 0) {
             stepperFeedbackTypeComposite.addComponent(new DisabledBottomNavigationStepperFeedbackType(stepperLayout));
+        }
+
+        if ((feedbackTypeMask & StepperFeedbackType.DISABLED_CONTENT_INTERACTION) != 0) {
+            stepperFeedbackTypeComposite.addComponent(new DisabledContentInteractionStepperFeedbackType(stepperLayout));
         }
 
         return stepperFeedbackTypeComposite;