+++ material-stepper/src/main/java/com/stepstone/stepper/internal/widget/StepViewPager.java
@@ -33,6 +33,8 @@ import static android.support.annotation.RestrictTo.Scope.LIBRARY;
 @RestrictTo(LIBRARY)
 public class StepViewPager extends ViewPager {
 
+    private boolean mBlockTouchEventsFromChildrenEnabled;
+
     public StepViewPager(Context context) {
         this(context, null);
     }
@@ -45,13 +47,19 @@ public class StepViewPager extends ViewPager {
     @Override
     public boolean onInterceptTouchEvent(MotionEvent event) {
         // Never allow swiping to switch between pages
-        return false;
+        return mBlockTouchEventsFromChildrenEnabled;
     }
 
     @Override
     public boolean onTouchEvent(MotionEvent event) {
         // Never allow swiping to switch between pages
-        return false;
+        return mBlockTouchEventsFromChildrenEnabled;
     }
 
+    /**
+     * @param blockTouchEventsFromChildrenEnabled true if children should not receive touch events
+     */
+    public void setBlockTouchEventsFromChildrenEnabled(boolean blockTouchEventsFromChildrenEnabled) {
+        this.mBlockTouchEventsFromChildrenEnabled = blockTouchEventsFromChildrenEnabled;
+    }
 }